Extract GitHub panel to separate component
