+++ packages/client/components/TeamPrompt/TeamPromptWorkDrawer.tsx
@@ -4,13 +4,12 @@ import React, {useState} from 'react'
 import {PreloadedQuery, useFragment, usePreloadedQuery} from 'react-relay'
 import {TeamPromptWorkDrawerQuery} from '../../__generated__/TeamPromptWorkDrawerQuery.graphql'
 import {TeamPromptWorkDrawer_meeting$key} from '../../__generated__/TeamPromptWorkDrawer_meeting.graphql'
-import clsx from 'clsx'
 import Tabs from '../Tabs/Tabs'
 import Tab from '../Tab/Tab'
 import ParabolLogoSVG from '../ParabolLogoSVG'
 import GitHubSVG from '../GitHubSVG'
-import GitHubObjectCard from './WorkDrawer/GitHubObjectCard'
 import ParabolTasksPanel from './WorkDrawer/ParabolTasksPanel'
+import GitHubIntegrationPanel from './WorkDrawer/GitHubIntegrationPanel'
 
 interface Props {
   queryRef: PreloadedQuery<TeamPromptWorkDrawerQuery>
@@ -25,83 +24,14 @@ const TeamPromptWorkDrawer = (props: Props) => {
       query TeamPromptWorkDrawerQuery($after: DateTime, $teamId: ID!, $userIds: [ID!]) {
         viewer {
           id
-          teamMember(teamId: $teamId) {
-            integrations {
-              github {
-                isActive
-                api {
-                  errors {
-                    message
-                    locations {
-                      line
-                      column
-                    }
-                    path
-                  }
-                  query {
-                    issues: search(
-                      first: 15
-                      type: ISSUE
-                      query: "is:issue sort:updated assignee:@me"
-                    ) @connection(key: "TeamPromptMeeting_issues") {
-                      edges {
-                        node {
-                          __typename
-                          ... on _xGitHubIssue {
-                            id
-                            title
-                            number
-                            repository {
-                              nameWithOwner
-                            }
-                            url
-                            state
-                            updatedAt
-                            lastEvent: timelineItems(last: 1) {
-                              updatedAt
-                            }
-                          }
-                        }
-                      }
-                    }
-                    pullRequests: search(
-                      first: 15
-                      type: ISSUE
-                      query: "is:pr sort:updated author:@me"
-                    ) @connection(key: "TeamPromptMeeting_pullRequests") {
-                      edges {
-                        node {
-                          __typename
-                          ... on _xGitHubPullRequest {
-                            id
-                            title
-                            number
-                            repository {
-                              nameWithOwner
-                            }
-                            url
-                            state
-                            updatedAt
-                            lastEvent: timelineItems(last: 1) {
-                              updatedAt
-                            }
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
           ...ParabolTasksPanel_user @arguments(userIds: $userIds)
+          ...GitHubIntegrationPanel_user @arguments(teamId: $teamId)
         }
       }
     `,
     queryRef
   )
   const {viewer} = data
-  const {teamMember} = viewer
 
   const meeting = useFragment(
     graphql`
@@ -112,18 +42,7 @@ const TeamPromptWorkDrawer = (props: Props) => {
     meetingRef
   )
 
-  const github = teamMember?.integrations.github
-
-  const githubIssues = teamMember?.integrations.github?.api?.query?.issues.edges?.map(
-    (edge) => edge?.node
-  )
-  const githubPullRequests = teamMember?.integrations.github?.api?.query?.pullRequests.edges?.map(
-    (edge) => edge?.node
-  )
-
   const [activeIdx, setActiveIdx] = useState(0)
-  const [githubType, setGithubType] = useState<'issue' | 'pr'>('issue')
-  const githubObjects = githubType === 'issue' ? githubIssues : githubPullRequests
 
   return (
     <>
@@ -163,63 +82,7 @@ const TeamPromptWorkDrawer = (props: Props) => {
       {activeIdx === 0 ? (
         <ParabolTasksPanel userRef={viewer} meetingRef={meeting} />
       ) : (
-        <>
-          {github?.isActive ? (
-            <>
-              <div className='my-4 flex w-full gap-2 px-4'>
-                <div
-                  key={'issue'}
-                  className={clsx(
-                    'grow cursor-pointer rounded-full py-2 px-4 text-center text-sm leading-3 text-slate-800',
-                    'issue' === githubType
-                      ? 'bg-grape-700 font-semibold text-white focus:text-white'
-                      : 'border border-slate-300 bg-white'
-                  )}
-                  onClick={() => setGithubType('issue')}
-                >
-                  Your issues
-                </div>
-                <div
-                  key={'pr'}
-                  className={clsx(
-                    'grow cursor-pointer rounded-full py-2 px-4 text-center text-sm leading-3 text-slate-800',
-                    'pr' === githubType
-                      ? 'bg-grape-700 font-semibold text-white focus:text-white'
-                      : 'border border-slate-300 bg-white'
-                  )}
-                  onClick={() => setGithubType('pr')}
-                >
-                  Your pull requests
-                </div>
-              </div>
-              <div className='flex flex h-full flex-col gap-y-2 overflow-auto px-4'>
-                {githubObjects?.map((object, idx) => {
-                  if (
-                    object?.__typename === '_xGitHubIssue' ||
-                    object?.__typename === '_xGitHubPullRequest'
-                  ) {
-                    return (
-                      <GitHubObjectCard
-                        type={githubType}
-                        key={idx}
-                        title={object.title}
-                        status={object.state}
-                        number={object.number}
-                        repoName={object.repository.nameWithOwner}
-                        url={object.url}
-                        updatedAt={object.lastEvent.updatedAt}
-                      />
-                    )
-                  } else {
-                    return null
-                  }
-                })}
-              </div>
-            </>
-          ) : (
-            <div>needs auth</div>
-          )}
-        </>
+        <GitHubIntegrationPanel userRef={viewer} />
       )}
     </>
   )
